requirements.txt
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.21.0
dnspython>=2.3.0 